feat: Use .cmd and .ps1 files on Windows
symlinks and junctions do not work

BREAKING CHANGE: Although the previous version did not work on Windows
in my tests, for somebdy it might. The new version creates .cmd and .ps1
files instead of the links. The junction will not exist in .bin after
installing a package. Otherwise the functionality did not change. The
binary can be run just by the file name with no file extension.